Lower limit when calling `ncdc_stations`
But check against the meta data to make sure that the function still pulls all the available data for each county. I'm hoping that this might help out a bit with the daily limit thing (although, I don't know-- if you set a high limit but don't use it, does it only count what you use? We should check, because the function would be more efficient I think if we left it as-is if there's no penalty for having a high limit if you only end up using a small amount of it).