export daily_stations
Hello, this function is not currently exported in the NAMESPACE. I think the intent is for it to be exported considering it is documented, and used in the examples for other function. I'm just making this edit through the github webbrowser, so you'll need to rerun roxygen document after accepting. 

Thanks for the package, i'm just trying it out right now but it seems pretty useful. I'm using ::: to access this function for now.