Update thresholds for unreasonable data
Some of the missing values in the weather data are coded with a series of "9"s, so they could slip through in some cases. We have thresholds to filter out unreasonable values. These were pretty close to the record values, so they might have accidently filtered some extreme but real values. This moves those thresholds out some, to leave room for changing extremes.